// src/app/core/models/ms-convertiongood/asset-conversion.ts
export interface IAssetConversion {
  goodId: string | number;
  conversionId: string | number;
  description: string;
  amount: number;
  status: string;
  val1: string;
  val2: string;
  val3: string;
  val4: string;
  val5: string;
  val6: string;
  val7: string;
  val8: string;
  val9: string;
  val10: string;
  val11: string;
  val12: string;
  val13: string;
  val14: string;
  val15: string;
  val16: string;
  val17: string;
  val18: string;
  val19: string;
  val20: string;
  val21: string;
  val22: string;
  val23: string;
  val24: string;
  val25: string;
  val26: string;
  val27: string;
  val28: string;
  val29: string;
  val30: string;
  val31: string;
  val32: string;
  val33: string;
  val34: string;
  val35: string;
  val36: string;
  val37: string;
  val38: string;
  val39: string;
  val40: string;
  val41: string;
  val42: string;
  val43: string;
  val44: string;
  val45: string;
  val46: string;
  val47: string;
  val48: string;
  val49: string;
  val50: string;
  appraisedValue: string | number;
  noClassifGood: string | number;
  noLabel: string | number;
  unit: string | number;
  noRegistration: string | number;
  descriptionConv: string;
  amountConv: string | number;
  noClasifGoodConv: string | number;
  unitConv: string | number;
  noLabelConv: string | number;
}